// ui/src/backend/generated/SessionState.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { DatasetSummary } from "./DatasetSummary";
import type { LoadingInfo } from "./LoadingInfo";
import type { PlotConfig } from "./PlotConfig";

export type SessionState = { dataset: DatasetSummary | null, plot: PlotConfig | null, loading: LoadingInfo | null, };